backend_kanban/pyproject.toml
[tool.poetry]
name = "backend-kanban"
version = "0.1.0"
description = ""
authors = ["eduardolirainfo <eduardolirainfo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
tortoise-orm = { extras = ["asyncpg"], version = "^0.20.0" }
psycopg2-binary = "^2.9.9"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"
pydantic = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
